ml_service/requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
psycopg2-binary
numpy
Pillow
python-multipart
python-dotenv
pyyaml
